@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -13,47 +13,48 @@ dependencies:
   # MobX is a library for reactively managing the state of your applications.
   # Use the power of observables, actions, and reactions to supercharge your Dart and Flutter apps.
   # See the following link for more information: https://pub.dev/packages/mobx
-  mobx: 2.0.5
+  mobx: ^2.0.7
 
   # Flutter integration for MobX. It provides a set of Observer widgets
   # that automatically rebuild when the tracked observables change.
   # See the following link for more information: https://pub.dev/packages/flutter_mobx
-  flutter_mobx: 2.0.2
+  flutter_mobx: ^2.0.6
 
   # A wrapper around InheritedWidget to make them easier to use and more reusable.
   # See the following link for more information: https://pub.dev/packages/provider
-  provider: 6.0.1
+  provider: ^6.0.3
 
   # AutoRoute is a declarative routing solution, where everything needed
   # for navigation is automatically generated for you.
   # See the following link for more information: https://pub.dev/packages/auto_route
-  auto_route: 3.2.0
+  auto_route: ^4.2.1
 
   # Implementations of SHA, MD5, and HMAC cryptographic functions
   # See the following link for more information: https://pub.dev/packages/crypto
-  crypto: 3.0.1
+  crypto: ^3.0.2
 
   # Flutter plugin for reading and writing simple key-value pairs.
   # Wraps NSUserDefaults on iOS and SharedPreferences on Android.
   # See the following link for more information: https://pub.dev/packages/shared_preferences
-  shared_preferences: 2.0.11
+  shared_preferences: ^2.0.15
 
   # A package that allows you to use a native file explorer to pick single
   # or multiple absolute file paths, with extension filtering support.
   # See the following link for more information: https://pub.dev/packages/file_picker
-  file_picker: 4.2.7
+  file_picker: ^5.0.1
 
   # Dialogs for picking and saving files in Android and in iOS.
-  flutter_file_dialog: 2.3.0
+  # See the following link for more information: https://pub.dev/packages/flutter_file_dialog
+  flutter_file_dialog: ^2.3.0
 
   # Flutter plugin for querying information about the application package,
   # such as CFBundleVersion on iOS or versionCode on Android.
   # See the following link for more information: https://pub.dev/packages/package_info_plus
-  package_info_plus: 1.3.0
+  package_info_plus: ^1.4.3
 
   # Flutter plugin for launching a URL. Supports web, phone, SMS, and email schemes.
   # See the following link for more information: https://pub.dev/packages/url_launcher
-  url_launcher: 6.0.17
+  url_launcher: ^6.1.5
 
 dev_dependencies:
   flutter_test:
@@ -61,16 +62,16 @@ dev_dependencies:
 
   # Recommended lints for Flutter apps, packages, and plugins to encourage good coding practices.
   # See the following link for more information: https://pub.dev/packages/flutter_lints
-  flutter_lints: 1.0.0
+  flutter_lints: ^2.0.1
 
   # A build system for Dart code generation and modular compilation.
   # See the following link for more information: https://pub.dev/packages/build_runner
-  build_runner: 2.1.5
+  build_runner: ^2.2.0
 
   # Code generator for MobX that adds support for annotating your code
   # with @observable, @computed, @action and also creating Store classes.
   # See the following link for more information: https://pub.dev/packages/mobx_codegen
-  mobx_codegen: 2.0.4
+  mobx_codegen: ^2.0.7
 
   # AutoRoute is a declartive routing solution, where everything needed
   # for navigation is automatically generated for you.
