@@ -10,7 +10,8 @@ import 'package:provider/provider.dart';
 void main() => runApp(const AppDependencies(app: App()));
 
 class App extends StatefulWidget {
-  const App({Key? key}) : super(key: key);
+
+  const App({super.key});
 
   @override
   State<App> createState() => _AppState();
@@ -22,7 +23,7 @@ class _AppState extends State<App> {
   @override
   void initState() {
     super.initState();
-    final widgetBindings = WidgetsBinding.instance!;
+    final widgetBindings = WidgetsBinding.instance;
     widgetBindings.window.onPlatformBrightnessChanged = () {
       if (context.read<SettingsRepository>().currentTheme() == AppTheme.system) {
         widgetBindings.handlePlatformBrightnessChanged();
